Rename getPageMarkdownData to getFrontmatter
